scripts/rag_data_loader_requirements.txt
sentence-transformers==3.0.1
chromadb
langchain_chroma
langchain_huggingface
langchain-community
pypdf 
python-dotenv